Extending certain prestates to include AosDrafted
