+++ definitions/divorce/json/CaseEvent/CaseEvent-nonprod.json
@@ -9,5 +9,17 @@
     "PreConditionState(s)": "AwaitingServicePayment",
     "PostConditionState": "AwaitingServiceConsideration",
     "SecurityClassification": "Public"
+  },
+  {
+    "LiveFrom": "26/02/2021",
+    "CaseTypeID": "DIVORCE",
+    "ID": "testAosDrafted",
+    "Name": "Aos Drafted Test",
+    "Description": "TEST ONLY: Aos Drafted Test",
+    "DisplayOrder": 20,
+    "PreConditionState(s)": "*",
+    "PostConditionState": "AosDrafted",
+    "RetriesTimeoutURLSubmittedEvent": "120,120",
+    "SecurityClassification": "Public"
   }
 ]
