+++ definitions/divorce/json/CaseEvent/CaseEvent-resp-journey-nonprod.json
@@ -35,5 +35,43 @@
     "CallBackURLAboutToSubmitEvent": "${CCD_DEF_COS_URL}/received-service-added-date",
     "RetriesTimeoutURLAboutToSubmitEvent": "120,120",
     "SecurityClassification": "Public"
+  },
+  {
+    "LiveFrom": "26/02/2021",
+    "CaseTypeID": "DIVORCE",
+    "ID": "aosNotReceived",
+    "Name": "AOS Not Received within SLA",
+    "Description": "AOS Not Received within SLA",
+    "DisplayOrder": 3,
+    "PreConditionState(s)": "AosAwaiting;AosDrafted",
+    "PostConditionState": "AosOverdue",
+    "CallBackURLAboutToSubmitEvent": "${CCD_DEF_COS_URL}/prepare-aos-not-received-for-submission",
+    "CallBackURLSubmittedEvent": "${CCD_DEF_COS_URL}/petition-updated",
+    "RetriesTimeoutURLSubmittedEvent": "120,120",
+    "SecurityClassification": "Public"
+  },
+  {
+    "LiveFrom": "26/02/2021",
+    "CaseTypeID": "DIVORCE",
+    "ID": "aosNotReceivedForProcessServer",
+    "Name": "Move to awaiting DN",
+    "Description": "Move to awaiting DN",
+    "DisplayOrder": "20",
+    "PreConditionState(s)": "AosAwaiting;AosStarted;AosDrafted",
+    "PostConditionState": "AwaitingDecreeNisi",
+    "CallBackURLSubmittedEvent": "${CCD_DEF_COS_URL}/aos-not-received-for-process-server",
+    "SecurityClassification": "Public"
+  },
+  {
+    "LiveFrom": "26/02/2021",
+    "CaseTypeID": "DIVORCE",
+    "ID": "aosNotReceivedForAltMethod",
+    "Name": "Move to awaiting DN",
+    "Description": "Move to awaiting DN",
+    "DisplayOrder": 26,
+    "PreConditionState(s)": "AosAwaiting;AosStarted;AosDrafted",
+    "PostConditionState": "AwaitingDecreeNisi",
+    "CallBackURLSubmittedEvent": "${CCD_DEF_COS_URL}/alternative-service-confirmed",
+    "SecurityClassification": "Public"
   }
 ]
