+++ test/unit/utils/utils.js
@@ -72,6 +72,31 @@ function getUniqValues(objectArray, property) {
   });
 }
 
+function byCaseType(caseType) {
+  return entry => {
+    return entry.CaseTypeID === caseType;
+  };
+}
+
+function byStateName(stateEntry) {
+  return stateAuth => {
+    return stateAuth.CaseStateID === stateEntry.ID;
+  };
+}
+
+function mapErrorArray(caseType) {
+  return entry => {
+    return {
+      UserRole: entry.UserRole,
+      CaseType: caseType
+    };
+  };
+}
+
+function missingAuthorisationsExist(missingAuthCount) {
+  return missingAuthCount > 0;
+}
+
 module.exports = {
   SHORT_STRING,
   MEDIUM_STRING,
@@ -84,5 +109,9 @@ module.exports = {
   isNotLongerThan,
   isPositiveNumber,
   whenPopulated,
-  getUniqValues
+  getUniqValues,
+  byCaseType,
+  byStateName,
+  mapErrorArray,
+  missingAuthorisationsExist
 };
